+++ main/app/layout.tsx
@@ -1,0 +1,35 @@
+import type { Metadata } from 'next';
+import '../styles/global.css';
+import { Providers } from './providers';
+import Header from '../src/views/Header/Header';
+import ErrorBoundary from '../src/components/ErrorBoundary/ErrorBoundary';
+
+export const metadata: Metadata = {
+  title: 'pokedex by @sergiozeppo',
+  description:
+    'React, Typescript, React Router, Next.js (Pages API / Router API), Vitest',
+  icons: {
+    icon: 'assets/img/favicon.png',
+  },
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <body>
+        <Providers>
+          <div className="container">
+            <ErrorBoundary>
+              <Header />
+              {children}
+            </ErrorBoundary>
+          </div>
+        </Providers>
+      </body>
+    </html>
+  );
+}
