+++ main/src/components/CardList/CardList.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import { useSelector } from 'react-redux';
 import { RootState } from '../../store';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import Card from '../Card/Card';
 import PokeLoader from '../PokeLoader/PokeLoader';
 import styles from './CardList.module.css';
