+++ main/src/components/ErrorBoundary/ErrorBoundary.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Component, ReactNode, ErrorInfo } from 'react';
 import styles from './ErrorBoundary.module.css';
