+++ main/src/components/ThemeSwitcher/ThemeSwitcher.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useContext, useEffect, useState } from 'react';
 import { ThemeContext } from '../../context/ThemeContext';
