+++ main/src/context/ThemeContext.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { createContext } from 'react';
 
 export type Theme = 'light' | 'dark';
