+++ main/src/context/ThemeProvider.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 import { Theme, ThemeContext } from './ThemeContext';
